pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deta"
version = "0.2.1"
description = "Infrastructure anomaly detection and monitoring tool"
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
authors = [
    {name = "Tom Sapletta", email = "tom@sapletta.com"},
]
keywords = ["infrastructure", "monitoring", "docker", "devops", "anomaly-detection"]
dependencies = [
    "ruamel.yaml>=0.17.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.24.0",
    "watchfiles>=0.20.0",
    "networkx>=3.0",
    "rich>=13.0",
    "typer>=0.9.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Quality Assurance",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/semcod/deta"
Repository = "https://github.com/semcod/deta"

[project.optional-dependencies]
docker = ["docker>=6.0"]
toml = ["tomli>=2.0; python_version<'3.11'"]

[project.scripts]
deta = "deta.cli:main"

[tool.setuptools.packages.find]
include = ["deta*"]

[tool.pfix]
# Self-healing Python configuration
model = "openrouter/qwen/qwen3-coder-next"
auto_apply = true
auto_install_deps = true
auto_restart = false
max_retries = 3
create_backups = false
git_auto_commit = false

[tool.pfix.runtime_todo]
enabled = true
todo_file = "TODO.md"
min_severity = "low"
deduplicate = true

[tool.costs]
# AI Cost tracking configuration
badge = true
update_readme = true
readme_path = "README.md"
default_model = "openrouter/qwen/qwen3-coder-next"
analysis_mode = "byok"
full_history = true
max_commits = 500

# Cost thresholds for badge colors (USD)
badge_color_thresholds = { low = 1.0, medium = 5.0, high = 10.0, critical = 50.0 }
